@@ -14,29 +14,29 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "chalk": "^2.4.2",
+    "chalk": "^4.1.2",
     "demux": "^4.0.0",
     "demux-eos": "^4.0.1",
-    "eosjs": "^20.0.3",
+    "eosjs": "^22.1.0",
     "human-to-cron": "^0.3.1",
     "lodash.get": "^4.4.2",
-    "massive": "^5.7.4",
-    "node-cron": "^2.0.3",
-    "node-fetch": "^2.3.0"
+    "massive": "^6.10.0",
+    "node-cron": "^3.0.0",
+    "node-fetch": "^3.0.0"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-config-standard-react": "^9.1.0",
-    "eslint-plugin-node": "^9.2.0",
-    "eslint-plugin-standard": "^4.0.1",
-    "husky": "^3.0.4",
-    "lint-staged": "^9.2.5",
-    "prettier": "^1.18.2",
-    "pretty-quick": "^1.11.1",
-    "react-scripts": "3.1.1",
-    "snazzy": "^8.0.0",
-    "standard": "^14.1.0"
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-config-standard-react": "^11.0.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-standard": "^5.0.0",
+    "husky": "^7.0.2",
+    "lint-staged": "^11.1.2",
+    "prettier": "^2.3.2",
+    "pretty-quick": "^3.1.1",
+    "react-scripts": "4.0.3",
+    "snazzy": "^9.0.0",
+    "standard": "^16.0.3"
   },
   "lint-staged": {
     "*.js": [
