@@ -1,12 +1,12 @@
-import React from 'react'
-import SvgIcon from '@material-ui/core/SvgIcon'
-import { makeStyles } from '@material-ui/core/styles'
+import React, { memo } from 'react'
+import SvgIcon from '@mui/material/SvgIcon'
+import { makeStyles } from '@mui/styles'
 
 import styles from './styles'
 
 const useStyles = makeStyles(styles)
 
-const DevelopmentIcon = (props) => {
+const DevelopmentIcon = props => {
   const classes = useStyles()
 
   return (
@@ -22,4 +22,4 @@ const DevelopmentIcon = (props) => {
   )
 }
 
-export default DevelopmentIcon
+export default memo(DevelopmentIcon)
