@@ -1,7 +1,9 @@
 import en from './en.json'
 import es from './es.json'
+import ko from './ko.json'
 
 export default {
   en,
-  es
+  es,
+  ko
 }
