@@ -1,7 +1,9 @@
-export default (theme) => ({
+export default theme => ({
   wrapperContainers: {
     color: '#433F5B',
     padding: theme.spacing(6, 2),
+    maxWidth: 1024,
+    width: '100%',
     '& h6': {
       marginBottom: theme.spacing(1)
     },
@@ -10,6 +12,8 @@ export default (theme) => ({
     }
   },
   firstContainer: {
+    display: 'flex',
+    justifyContent: 'center',
     backgroundColor: theme.palette.surface.main,
     '& h5': {
       fontSize: theme.typography.h4.fontSize,
@@ -17,12 +21,16 @@ export default (theme) => ({
     }
   },
   middleContainer: {
+    display: 'flex',
+    justifyContent: 'center',
     backgroundColor: theme.palette.surface.main,
     '& img': {
       width: '100%'
     }
   },
   lastContainer: {
+    display: 'flex',
+    justifyContent: 'center',
     backgroundColor: '#f5f5f5'
   }
 })
